+++ source/exceeds_expectations/package.d
@@ -1,4 +1,4 @@
 module exceeds_expectations;
 
-public import exceeds_expectations.expectation;
+public import exceeds_expectations.expect;
 public import exceeds_expectations.expect_not;